+++ objectRecognitionSource/src/test/java/io/vantiq/extsrc/objectRecognition/neuralNet/TestYoloProcessor.java
@@ -112,45 +112,49 @@ public class TestYoloProcessor extends NeuralNetTestBase {
             fail("Could not setup the JSON YoloProcessor");
         }
 
-        vantiq = new io.vantiq.client.Vantiq(testVantiqServer);
-        vantiq.setAccessToken(testAuthToken);
+        if (testVantiqServer != null && testAuthToken != null) {
+            vantiq = new io.vantiq.client.Vantiq(testVantiqServer);
+            vantiq.setAccessToken(testAuthToken);
+        }
     }
 
     @AfterClass
     public static void deleteFromVantiq() throws InterruptedException {
         // Deleting files saved as documents
-        for (int i = 0; i < vantiqSavedFiles.size(); i++) {
-            Thread.sleep(1000);
-            vantiq.deleteOne(VANTIQ_DOCUMENTS, vantiqSavedFiles.get(i), new BaseResponseHandler() {
+        if (vantiq != null && vantiq.isAuthenticated()) {
+            for (int i = 0; i < vantiqSavedFiles.size(); i++) {
+                Thread.sleep(1000);
+                vantiq.deleteOne(VANTIQ_DOCUMENTS, vantiqSavedFiles.get(i), new BaseResponseHandler() {
 
-                @Override
-                public void onSuccess(Object body, Response response) {
-                    super.onSuccess(body, response);
-                }
+                    @Override
+                    public void onSuccess(Object body, Response response) {
+                        super.onSuccess(body, response);
+                    }
 
-                @Override
-                public void onError(List<VantiqError> errors, Response response) {
-                    super.onError(errors, response);
-                }
+                    @Override
+                    public void onError(List<VantiqError> errors, Response response) {
+                        super.onError(errors, response);
+                    }
 
-            });
-        }
-        // Deleting files saved as images
-        for (int i = 0; i < vantiqSavedImageFiles.size(); i++) {
-            Thread.sleep(1000);
-            vantiq.deleteOne(VANTIQ_IMAGES, vantiqSavedImageFiles.get(i), new BaseResponseHandler() {
+                });
+            }
+            // Deleting files saved as images
+            for (int i = 0; i < vantiqSavedImageFiles.size(); i++) {
+                Thread.sleep(1000);
+                vantiq.deleteOne(VANTIQ_IMAGES, vantiqSavedImageFiles.get(i), new BaseResponseHandler() {
 
-                @Override
-                public void onSuccess(Object body, Response response) {
-                    super.onSuccess(body, response);
-                }
+                    @Override
+                    public void onSuccess(Object body, Response response) {
+                        super.onSuccess(body, response);
+                    }
 
-                @Override
-                public void onError(List<VantiqError> errors, Response response) {
-                    super.onError(errors, response);
-                }
+                    @Override
+                    public void onError(List<VantiqError> errors, Response response) {
+                        super.onError(errors, response);
+                    }
 
-            });
+                });
+            }
         }
     }
 
@@ -166,10 +170,12 @@ public class TestYoloProcessor extends NeuralNetTestBase {
             deleteDirectory(OUTPUT_DIR);
         }
 
-        // Double check that everything was deleted from VANTIQ
-        deleteSource(vantiq);
-        deleteType(vantiq);
-        deleteRule(vantiq);
+        if (vantiq != null && vantiq.isAuthenticated()) {
+            // Double check that everything was deleted from VANTIQ
+            deleteSource(vantiq);
+            deleteType(vantiq);
+            deleteRule(vantiq);
+        }
     }
 
     @Test
